Follow file naming conventions to match module namespace
The files defining ImageSpec::Parser and nested modules/classes are
defined under lib/parser instead of lib/image_spec/parser. This can
cause issues with the "parser" gem, because depending on which one comes
first in the load path, a "require 'parser'" will load the wrong file.

This change makes the files follow the naming convention that is
established in the Ruby community.